paint_down_to_common: plug a memory leak
When a commit is missing, we return early (currently pretending that no
merge basis could be found in that case). At that stage, it is possible
that a merge base could have been found already, and added to the
`result`, which is now leaked.

Let's release it, to address that memory leak.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>